Scripts now ensure more reliable usb mounting. Log files are now copied to the usb.

+++ projects/camstreamer/scripts/start_usb_install.sh
@@ -3,8 +3,7 @@
 
 set -e # exit shell on a simple command failure
 install_dir=$1 # eg. /actioncamstreamer
-dev_name=$2 # eg. /dev/sdd
-partition_num=1
+dev_name=/dev/$2 # eg. /dev/sdd1
 mount_dir=/mount/usb_drive/
 usb_firmware_dir=$mount_dir/GoPie/
 binary_name=actioncamstreamer
@@ -12,44 +11,49 @@ binary_name=actioncamstreamer
 # start LED flash
 led_dir=/sys/class/leds/led0
 echo "timer" >> $led_dir/trigger
-echo 250 >> $led_dir/delay_off
-echo 250 >> $led_dir/delay_on
+echo 150 >> $led_dir/delay_off
+echo 150 >> $led_dir/delay_on
 
 # mount the drive
 mkdir -p $mount_dir
-mount $dev_name$partition_num $mount_dir
+mount $dev_name $mount_dir
 
-# install
-# run install here for safety, instead of running a script on the usb
 cd $install_dir
-cp $usb_firmware_dir/firmware.tgz . # copy the compressed firmware install files
-tar xzf firmware.tgz # extract the files
 
 # First kill the firmware and start script
 # This is killed first because it can control the LED
 systemctl kill --kill-who=all start_actioncamstreamer.service
 
-# copy the new firmware
-cp firmware/* .
-chmod +x $binary_name
+# install firmware if it exists on the usb
+# run install here for safety, instead of running a script on the usb
+if [ -e "$usb_firmware_dir/firmware.tgz" ]
+then
+    cp $usb_firmware_dir/firmware.tgz . # copy the compressed firmware install files
+    tar xzf firmware.tgz # extract the files
 
-# copy recent logs to the usb
-config_filename="GoPie_userconfig.txt"
-log_filename="GoPie_logs.txt"
+    # copy the new firmware
+    cp firmware/* .
+    chmod +x $binary_name
+fi
 
 # copy log file if it exists
-if [ -f "$log_filename" ]
+if [ -e *.log ]
 then
-  cp $log_filename $usb_firmware_dir
+    cp *.log $usb_firmware_dir
+    if [ -d backup_logs ]
+    then
+	cp -r backup_logs $usb_firmware_dir
+    fi
 fi
 
 # copy config file if it exists
-if [ -f "$usb_firmware_dir/$config_filename" ]
+if [ -e "$usb_firmware_dir/GoPie_userconfig.txt" ]
 then
   cp $usb_firmware_dir/$config_filename .
 fi
 
 sleep 3 # so that the user has time to see that the install has triggered
+echo "0" >> $led_dir/brightness # turn off
 
 # Restart the firmware start script service. The firmware will then take control of the led.
 systemctl restart start_actioncamstreamer.service
